remove redundant next auth handler
